+++ src/components/CardNumber/index.jsx
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
-import FieldSet from '../../FieldSet';
-import CardNumberInput from '../../Input/CardNumberInput';
+import FieldSet from '../FieldSet';
+import CardNumberInput from '../Input/CardNumberInput';
 
 const CardNumber = ({
   firstCardNumber,
